requirements.txt
# 核心依赖
python-dotenv>=0.19.0
pydantic>=1.8.2
typing-extensions>=4.0.0

# AI相关
openai>=0.27.0
langchain>=0.0.200
# faiss-cpu>=1.7.4  # 向量数据库 - 暂时注释，需要swig依赖

# 代码处理
astroid>=2.9.3
pylint>=2.12.2
black>=21.12b0

# 测试相关
pytest>=6.2.5
pytest-cov>=3.0.0

# 版本控制
gitpython>=3.1.24

# 工具
tqdm>=4.62.3
loguru>=0.5.3
